requirements.txt
dissect.cstruct==3.2